PYG 3 PYG 34 (feat): Lógica para Validade e Salvar o Agendador
